peltool: Handle bad JSON in from UD parsers
The UserData parsers are supposed to return a string that contains valid
JSON.  If they return invalid JSON then peltool will crash with a JSON
exception.

Fix that by catching the JSON exception and printing a hexdump of the
data instead.

This was happening when a UserData section that contained JSON (the
built in JSON created by the BMC) was truncated in a large PEL so wasn't
valid.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>