peltool: Add -O option
This commit introduces the -O/--only option.
Using this option we can add exclusivity to options.
We consider only pels that match the selected options

Tested on sample PELs:

In below example: by default we have 24 pels, When we include hidden, we have 28 pels. When we select only hidden, we can see there are 4 pels.
```bash
$ peltool.py -p testpels/ -n
{
    "Number of PELs found": 24
}
$ peltool.py -p testpels/ -nH
{
    "Number of PELs found": 28
}
$ peltool.py -p testpels/ -nOH
{
    "Number of PELs found": 4
}
```
In below example: by default It shows 24 pels, as Predictive pels are included in it. If we want to see only Predictive, the count is 17.
Similarly number of pels including Predictive and Informational is 46, and only Predictive and Informational is 39.
```bash
$ peltool.py -p testpels/ -nS Predictive
{
    "Number of PELs found": 24
}
$ peltool.py -p testpels/ -nOS Predictive
{
    "Number of PELs found": 17
}

$ peltool.py -p testpels/ -nS Predictive Informational
{
    "Number of PELs found": 46
}
$ peltool.py -p testpels/ -nOS Predictive Informational
{
    "Number of PELs found": 39
}

```

Signed-off-by: Harsh Agarwal <Harsh.Agarwal@ibm.com>